+++ test/contracts/note.test.ts
@@ -24,4 +24,39 @@ describe('should post note', () => {
     const { data: note } = await contract.getNote(profileId, data.noteId)
     expect(note.metadata?.title).toBe('test')
   })
+
+  test('mintNote', async () => {
+    const { data } = await contract.postNote(profileId, {
+      title: 'test',
+      summary: 'test',
+    })
+
+    expect(data.noteId).toBeDefined()
+
+    const { transactionHash: mintHash } = await contract.mintNote(
+      profileId,
+      data.noteId,
+      mockUser.address,
+    )
+
+    expect(mintHash).toBeDefined()
+
+    // TODO: how do i check the NFT?
+  })
+
+  test('deleteNote', async () => {
+    const { data } = await contract.postNote(profileId, {
+      title: 'test',
+      summary: 'test',
+    })
+
+    expect(data.noteId).toBeDefined()
+
+    const { transactionHash: deleteHash } = await contract.deleteNote(
+      profileId,
+      data.noteId,
+    )
+
+    expect(deleteHash).toBeDefined()
+  })
 })